+++ jgltf-model/src/main/java/de/javagl/jgltf/model/CameraModel.java
@@ -65,7 +65,7 @@ public interface CameraModel extends NamedModelElement
      * camera will be used.
      * @return The result array
      */
-    float[] computeProjectionMatrix(float result[], Float aspectRatio);
+    double[] computeProjectionMatrix(double result[], Float aspectRatio);
 
     /**
      * Create the supplier of the projection matrix for this camera model.<br>
@@ -84,7 +84,7 @@ public interface CameraModel extends NamedModelElement
      * aspect ratio of the camera will be used.
      * @return The supplier
      */
-    Supplier<float[]> createProjectionMatrixSupplier(
+    Supplier<double[]> createProjectionMatrixSupplier(
         DoubleSupplier aspectRatioSupplier);
 
 }
